src/downloaders/investing_downloader.py
# src/downloaders/investing_downloader.py

import pandas as pd
import requests
import logging
from bs4 import BeautifulSoup

class InvestingDownloader:
    """
    Classe pour télécharger des données depuis Investing.com.
    """

    def __init__(self):
        self.logger = logging.getLogger(self.__class__.__name__)

    def download(self, header, ticker, freq='M', offset=None):
        """
        Télécharge les données depuis Investing.com.

        Parameters
        ----------
        header : str
            Le nom de la colonne à utiliser dans le DataFrame.
        ticker : str
            Le ticker Investing.com à télécharger.
        freq : str, optional
            La fréquence des données ('M', etc.).
        offset : int, optional
            Offset à appliquer sur les dates.

        Returns
        -------
        pd.DataFrame
            DataFrame contenant les données téléchargées avec les colonnes 'Date' et le header spécifié.
        """
        try:
            # Exemple fictif: Remplacer par le code réel pour télécharger les données depuis Investing.com
            url = f"https://www.investing.com/indices/{ticker}-historical-data"
            headers = {'User-Agent': 'Mozilla/5.0'}
            response = requests.get(url, headers=headers)
            if response.status_code == 200:
                soup = BeautifulSoup(response.text, 'html.parser')
                # Parser les données historiques ici
                # Ceci est un exemple simplifié
                data = {
                    'Date': pd.date_range(end=pd.Timestamp.now(), periods=10, freq='M'),
                    header: pd.np.random.rand(10) * 100  # Utilisation de données aléatoires
                }
                df = pd.DataFrame(data)
                return df
            else:
                self.logger.error(f"Erreur lors du téléchargement des données Investing.com pour {ticker}: Status Code {response.status_code}")
                return pd.DataFrame()
        except Exception as e:
            self.logger.error(f"Erreur lors du téléchargement des données Investing.com pour {ticker}: {e}")
            return pd.DataFrame()