requirements.txt
pandas
pandas-datareader
requests
PyGithub
yfinance
git+https://github.com/rongardF/tvdatafeed.git